Move functional scripts to cli/
