+++ laser_setup/cli/__init__.py
@@ -1,0 +1,1 @@
+from . import console, find_dp_script, setup_adapters
